cleanup performance tests
