+++ performance/performance.cpp
@@ -9,31 +9,33 @@
 #include <iostream>
 #include <stdexcept>
 
-#include <boost/context/execution_context.hpp>
+#include <boost/context/continuation.hpp>
 #include <boost/cstdint.hpp>
 #include <boost/program_options.hpp>
 
-#include "../bind_processor.hpp"
-#include "../clock.hpp"
-#include "../cycle.hpp"
+#include "bind_processor.hpp"
+#include "clock.hpp"
+#include "cycle.hpp"
 
 boost::uint64_t jobs = 1000;
 
-static boost::context::execution_context< void > foo( boost::context::execution_context< void > && ctx) {
+namespace ctx = boost::context;
+
+static ctx::continuation foo( ctx::continuation && c) {
     while ( true) {
-        ctx = ctx();
+        c = c.resume();
     }
-    return std::move( ctx);
+    return std::move( c);
 }
 
 duration_type measure_time() {
     // cache warum-up
-    boost::context::execution_context< void > ctx( foo);
-    ctx = ctx();
+    ctx::continuation c = ctx::callcc( foo);
+    c = c.resume();
 
     time_point_type start( clock_type::now() );
     for ( std::size_t i = 0; i < jobs; ++i) {
-        ctx = ctx();
+        c = c.resume();
     }
     duration_type total = clock_type::now() - start;
     total -= overhead_clock(); // overhead of measurement
@@ -46,13 +48,13 @@ duration_type measure_time() {
 #ifdef BOOST_CONTEXT_CYCLE
 cycle_type measure_cycles() {
     // cache warum-up
-    boost::context::fixedsize_stack alloc;
-    boost::context::execution_context< void > ctx( std::allocator_arg, alloc, foo);
-    ctx = ctx();
+    ctx::fixedsize_stack alloc;
+    ctx::continuation c = ctx::callcc( std::allocator_arg, alloc, foo);
+    c = c.resume();
 
     cycle_type start( cycles() );
     for ( std::size_t i = 0; i < jobs; ++i) {
-        ctx = ctx();
+        c = c.resume();
     }
     cycle_type total = cycles() - start;
     total -= overhead_cycle(); // overhead of measurement
@@ -87,10 +89,10 @@ int main( int argc, char * argv[]) {
         }
 
         boost::uint64_t res = measure_time().count();
-        std::cout << "execution_context: average of " << res << " nano seconds" << std::endl;
+        std::cout << "continuation: average of " << res << " nano seconds" << std::endl;
 #ifdef BOOST_CONTEXT_CYCLE
         res = measure_cycles();
-        std::cout << "execution_context: average of " << res << " cpu cycles" << std::endl;
+        std::cout << "continuation: average of " << res << " cpu cycles" << std::endl;
 #endif
 
         return EXIT_SUCCESS;
