--- performance/ucontext/Jamfile.v2
@@ -1,64 +1,0 @@
-
-#          Copyright Oliver Kowalke 2009.
-# Distributed under the Boost Software License, Version 1.0.
-#    (See accompanying file LICENSE_1_0.txt or copy at
-#          http://www.boost.org/LICENSE_1_0.txt)
-
-# For more information, see http://www.boost.org/
-
-import common ;
-import feature ;
-import indirect ;
-import modules ;
-import os ;
-import toolset ;
-
-project boost/context/performance/ucontext
-    : requirements
-      <library>/boost/chrono//boost_chrono
-      <library>/boost/context//boost_context
-      <library>/boost/program_options//boost_program_options
-      <linkflags>"-lrt" 
-      <link>static
-      <optimization>speed
-      <threading>multi
-      <variant>release
-      <cxxflags>-DBOOST_DISABLE_ASSERTS
-    ;
-
-alias sources
-   : ../bind_processor_aix.cpp
-   : <target-os>aix
-   ;
-
-alias sources
-   : ../bind_processor_freebsd.cpp
-   : <target-os>freebsd
-   ;
-
-alias sources
-   : ../bind_processor_hpux.cpp
-   : <target-os>hpux
-   ;
-
-alias sources
-   : ../bind_processor_linux.cpp
-   : <target-os>linux
-   ;
-
-alias sources
-   : ../bind_processor_solaris.cpp
-   : <target-os>solaris
-   ;
-
-alias sources
-   : ../bind_processor_windows.cpp
-   : <target-os>windows
-   ;
-
-explicit sources ;
-
-exe performance
-   : sources
-     performance.cpp
-   ;
